+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Event/Attempt sub-resource to allow retrieval of information on failed webhook event delivery attempts
+- get, query, page, delete and update function for Event resource.
 
 ## [0.1.0] - 2022-06-03
 ### Added